# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin+gorm 后台API",
        "title": "gin+gorm 测试swagger",
        "contact": {
            "name": "go-swagger文档",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "v1.0"
    },
    "host": "222.196.37.70:8080",
    "paths": {
...
        "/api/createGroup": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "群聊管理模块"
                ],
                "summary": "新增群聊",
                "parameters": [
                    {
                        "description": "群聊信息",
                        "name": "Mgroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.Mgroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/deleteGroup": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "群聊管理模块"
                ],
                "summary": "删除群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群聊id",
                        "name": "gid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/downloadfile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "消息模块"
                ],
                "summary": "文件下载",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file路径",
                        "name": "filepath",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/getFhistoryMsg": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "消息模块"
                ],
                "summary": "获取与的好友历史消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "好友id",
                        "name": "fid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getGhistoryMsg": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "消息模块"
                ],
                "summary": "获取群聊历史消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "群聊id",
                        "name": "gid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getGroupByGid": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "群聊管理模块"
                ],
                "summary": "使用Gid查询群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群id",
                        "name": "gid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getGroupByName": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "群聊管理模块"
                ],
                "summary": "群名模糊查询群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群名",
                        "name": "gname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/getwsconn": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "消息模块"
                ],
                "summary": "建立websocket连接",
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/updateGroup": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "群聊管理模块"
                ],
                "summary": "更新群聊",
                "parameters": [
                    {
                        "description": "群聊信息",
                        "name": "Mgroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.Mgroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/uploadfile": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "消息模块"
                ],
                "summary": "文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.Mgroup": {
            "type": "object",
            "required": [
                "description",
                "groupname",
                "uid"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 250,
                    "minLength": 10
                },
                "gid": {
                    "type": "integer"
                },
                "groupname": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 2
                },
                "uid": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}